test: comprehensive suites, pinned lint rules, and CI that exercises the journey
Tests (294 total, all hardware-free and deterministic):
- run manifest serialization and tolerance of foreign/missing fields
- run store: atomic writes, traversal-safe ids and artifact paths,
  interrupted-run reconciliation, cancellation sentinels
- event ordering and persistence, sink fan-out, callback isolation,
  truncated-line tolerance
- config validation per field, presets, mission store, loader errors
- insight generation asserted on the *reason* each finding appears, plus
  priority ordering and the release verdict
- comparison compatibility, coverage changes, Markdown/JSON exports
- path-safety guards (including symlink escape), upload limits
- signal analysis: rate, gaps, outliers, orientation error
- HTML escaping of device output and plugin-supplied metric names
- a full simulated mission end to end, plus every new CLI command

Lint/type config: ruff's rule selection is now pinned explicitly, so a new
ruff release cannot fail CI on unchanged code. Fixing the resulting
findings surfaced a real bug — Path used in an annotation without being
imported in eaiv.core.metrics.

Two defects the tests found and fixed:
- partially-declared metric metadata silently lost unit and direction;
  each field now falls back to inference independently
- a lone spike in an otherwise constant signal escaped outlier detection
  because the MAD was zero; falls back to the mean absolute deviation

The regression gate no longer fails on metrics a suite labelled `mock`: a
stand-in runtime's sub-microsecond timings swing by orders of magnitude
between runs, so gating on them is noise. They stay visible as
informational. Legacy reports, which declare no provenance, gate exactly
as before. This let CI drop from a 200% allowance to the realistic default.

CI now also runs doctor, validates every shipped config and preset, drives
the recorded pipeline through promote-then-gate, inspects runs via the
CLI, verifies the dashboard's data layer can load the output, and runs the
guided demo.

+++ src/eaiv/dashboard/ui/views/telemetry_lab.py
@@ -151,9 +151,7 @@ def render(workspace: Workspace) -> None:
                 icon=None,
             )
             return
-        chosen = st.selectbox(
-            "Capture", candidates, format_func=lambda p: str(p), key="tl_file"
-        )
+        chosen = st.selectbox("Capture", candidates, format_func=lambda p: str(p), key="tl_file")
         path = Path(chosen)
         try:
             frame = pd.read_csv(path)
@@ -206,11 +204,11 @@ def render(workspace: Workspace) -> None:
                         {
                             "At (s)": f"{start:.4f}",
                             "Gap (s)": f"{length:.4f}",
-                            "Missing samples (est.)": max(
-                                0, int(round(length / sampling.median_interval_s)) - 1
-                            )
-                            if sampling.median_interval_s
-                            else 0,
+                            "Missing samples (est.)": (
+                                max(0, round(length / sampling.median_interval_s) - 1)
+                                if sampling.median_interval_s
+                                else 0
+                            ),
                         }
                         for start, length in sampling.gaps[:50]
                     ]
@@ -230,9 +228,7 @@ def render(workspace: Workspace) -> None:
     if x_axis and len(frame) > 1:
         t_min, t_max = float(frame["t_s"].min()), float(frame["t_s"].max())
         if t_max > t_min:
-            window = st.slider(
-                "Time range (s)", t_min, t_max, (t_min, t_max), key="tl_range"
-            )
+            window = st.slider("Time range (s)", t_min, t_max, (t_min, t_max), key="tl_range")
             frame = frame[(frame["t_s"] >= window[0]) & (frame["t_s"] <= window[1])]
 
     for title in selected_groups:
@@ -252,9 +248,7 @@ def render(workspace: Workspace) -> None:
             "Outliers are flagged with a modified Z score above 3.5 (median-based, so a few "
             "extreme samples cannot hide behind their own effect on the mean)."
         )
-        inspect = st.selectbox(
-            "Inspect outliers in", [s.name for s in flagged], key="tl_outlier"
-        )
+        inspect = st.selectbox("Inspect outliers in", [s.name for s in flagged], key="tl_outlier")
         target = next(s for s in flagged if s.name == inspect)
         indices = target.outlier_indices[:200]
         st.dataframe(
